test for file size
