+++ Wire-iOS Tests/ZipFile/URLs+zipTestsTests.swift
@@ -1,0 +1,37 @@
+
+// Wire
+// Copyright (C) 2020 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import XCTest
+@testable import Wire
+
+final class URLs_zipTests: XCTestCase {
+
+    func testThatFileURLsCanBeZipped(){
+        // GIVEN
+        let urls = ["not_animated.gif", "0x0.pdf", "sample.pkpass"].map() {
+            urlForResource(inTestBundleNamed:$0)
+        }
+
+        // WHEN
+        let zipURL = urls.zipFiles(filename: "test.zip")
+
+        // THEN
+        XCTAssertNotNil(zipURL)
+        XCTAssertGreaterThan(try! zipURL!.fileSize()!, 0)
+    }
+}
